# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AirDnD API
  version: 0.1.0
  description: AirDnD 프론트엔드와 백엔드가 공유하는 API 계약입니다.
servers:
  - url: http://127.0.0.1:8080
    description: 로컬 백엔드 서버
paths:
...
  /api/rooms/{roomId}:
    get:
      tags: [Rooms]
      summary: 숙소 상세 조회
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: 숙소 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetail"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  parameters:
    RoomId:
      name: roomId
      in: path
      required: true
      schema:
        type: integer
  responses:
...
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    RoomSummary:
      type: object
      required:
        [id, name, region, address, pricePerNight, rating, reviewCount, maxGuests, imageUrl, isAvailable, allowsPets]
      properties:
        id:
          type: integer
        name:
          type: string
        region:
          type: string
        address:
          type: string
        pricePerNight:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        maxGuests:
          type: integer
        imageUrl:
          type: string
          format: uri
        isAvailable:
          type: boolean
        allowsPets:
          type: boolean
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
          description: >
            지도 마커 표시용 위도. 지도 검색을 지원하려면 목록 응답에도 포함해야 합니다.
            (백엔드 RoomResponse가 추가하기 전까지는 누락될 수 있어 required 아님)
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
          description: 지도 마커 표시용 경도. latitude와 함께 제공되어야 합니다.
    RoomDetail:
      allOf:
        - $ref: "#/components/schemas/RoomSummary"
        - type: object
          required: [description, amenities, hostName, latitude, longitude]
          properties:
            description:
              type: string
            amenities:
              type: array
              items:
                type: string
            imageUrls:
              type: array
              items:
                type: string
                format: uri
            hostName:
              type: string
            latitude:
              type: number
            longitude:
              type: number
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        detail:
          type: object
          additionalProperties: true
...